Finished problem 2.
Created a city_state variable and a binary resolved variable of disposition. Fitted a logistic regression of resolved vs unresolved on victim age, sex and race for each city and extract the adjusted odds ratio (and CI) for solving homicides comparing male victims to female victims. Made a plot to show the estimates.